pyproject.toml
[tool.poetry]
name = "auth-service-fast-api"
version = "0.1.0"
description = ""
authors = ["medlcago"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.36"
email-validator = "^2.2.0"
asyncpg = "^0.30.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
python-multipart = "^0.0.16"
alembic = "^1.13.3"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.2" }
jinja2 = "^3.1.4"
dependency-injector = "^4.42.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
